requirements.txt
fastapi
uvicorn
slack_sdk
google-auth
google-auth-oauthlib
google-api-python-client
openai
python-dotenv